+++ ai/src/main/java/me/rerere/ai/provider/providers/openai/ChatCompletionsAPI.kt
@@ -1,3 +1,4 @@
+@file:Suppress("UNNECESSARY_SAFE_CALL")
 package me.rerere.ai.provider.providers.openai
 
 import android.util.Log