+++ app/src/main/java/net/weero/measix/pilot/ui/theme/Theme.kt
@@ -1,7 +1,7 @@
-﻿package net.weero.measix.pilot.ui.theme
+package net.weero.measix.pilot.ui.theme
 
-import android.app.Activity
 import android.os.Build
+import androidx.activity.compose.LocalActivity
 import androidx.compose.foundation.LocalOverscrollFactory
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialExpressiveTheme
@@ -78,13 +78,15 @@ fun MeasixTheme(
     val extendColors = if (darkTheme) ExtendDarkColors else ExtendLightColors
 
     // 更新状态栏图标颜色
+    val activity = LocalActivity.current
     val view = LocalView.current
     if (!view.isInEditMode) {
         SideEffect {
-            val window = (view.context as Activity).window
-            WindowCompat.getInsetsController(window, view).apply {
-                isAppearanceLightStatusBars = !darkTheme
-                isAppearanceLightNavigationBars = !darkTheme
+            activity?.window?.let { window ->
+                WindowCompat.getInsetsController(window, view).apply {
+                    isAppearanceLightStatusBars = !darkTheme
+                    isAppearanceLightNavigationBars = !darkTheme
+                }
             }
         }
     }